Dunno what, found this on my HD
You know, I really, *really* should have put this project in git when I
started it. Because now I keep finding different versions, some of which
are in git and some not, and the one on github previously seems
incomplete, and I basically have no idea what's correct. Probably nothing
is.